Created a new Docker environment for the "mots" service
A Dockerfile, Pipfile, and other necessary files were set up for the "mots" service. This includes setting up the Python environment and configuring Flask. The docker-compose file was also updated to include two new databases and the "mots" service. The databases are configured with healthcheck and exposed on their respective ports.