Add schema validation and refactor date formatting
A new package, `schemadict`, has been added to the Pipfile for schema validation. This package is used to create a schema for the MOT validator. Furthermore, the date format has been changed in the `MotModel` to ISO format in the `as_dict` method. Commits within the `update` method have also been included to persist data changes.