Update README and add MOT-related operations
This commit introduces an update to the README.md and README.polish.md, by adding details about the MOT-related operations. Also updated the wiki and coverage links for the 'mots' microservice in both README files. Furthermore, the 'coverage' and 'sphinx' entries in the Pipfile were edited, with 'sphinx-rtd-theme' added to the dependencies.