Add AWS-Resources microservice and adjust database volumes
This commit introduces a new microservice for managing AWS resources, and comprises new files for handling configuration, environment variables, and application setup as well as a new Dockerfile. Additionally, volumes for the 'mots' and 'cars' microservices in 'docker-compose.yml' have been correctly assigned. Spacing and formatting in comment sections of the file have also been adjusted for clarity.