Return full object instead of ID after creating Car and Mot
Instead of only returning the ID after a new Car and Mot object is created, this change returns the whole object. This modification affects both the Car and Mot routes, specifically the POST methods, and their corresponding tests were updated to reflect this change. Additionally, improved error logging was added in the Car route to handle ValueError or TypeError.