Add section on using `selenium` in Docker container to README
The new section in both English and Polish README versions details how to correctly configure `selenium` to work in a Docker container. It explains how to install Google Chrome in a Debian container, how to add necessary arguments to the `selenium.webdriver.chrome.options.Options` object, and finally, how to create a driver for proper functioning in a Docker container.