Set up Alembic and initial entities migration
This commit includes the introduction of Alembic for database migration management and the initial entities migration to create the `drivers` table. A series of configurations files are added for the Alembic environment, including `alembic.ini` and `env.py`. Initial entities migration add necessary columns to the `drivers` table. This commit also introduces the Dockerfile for deployment.